requirements.txt
pyTelegramBotAPI==4.10.0
aiohttp==3.8.4
python-dotenv